Remove PHP uopz extension dependency
