+++ .github/workflows/bc.yml
@@ -33,4 +33,3 @@ jobs:
         ['ubuntu-latest']
       php: >-
         ['8.3']
-      extensions: uopz
