+++ .github/workflows/composer-require-checker.yml
@@ -32,4 +32,3 @@ jobs:
         ['ubuntu-latest']
       php: >-
         ['8.1', '8.2', '8.3']
-      extensions: uopz
