+++ .github/workflows/rector.yml
@@ -22,4 +22,3 @@ jobs:
         ['ubuntu-latest']
       php: >-
         ['8.3']
-      extensions: uopz
