+++ .github/workflows/static.yml
@@ -29,4 +29,3 @@ jobs:
         ['ubuntu-latest']
       php: >-
         ['8.3']
-      extensions: uopz
