+++ composer.json
@@ -53,7 +53,6 @@
         "symfony/finder": "^5.4"
     },
     "require-dev": {
-        "ext-uopz": "*",
         "cycle/entity-behavior": "^1.0",
         "jetbrains/phpstorm-attributes": "^1.0",
         "maglnet/composer-require-checker": "^4.3",
@@ -77,10 +76,7 @@
     "autoload-dev": {
         "psr-4": {
             "Yiisoft\\Yii\\Cycle\\Tests\\": "tests"
-        },
-        "files": [
-            "tests/bootstrap.php"
-        ]
+        }
     },
     "extra": {
         "branch-alias": {
