--- tests/bootstrap.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * @link https://github.com/krakjoe/uopz/issues/172
- */
-if (function_exists('uopz_allow_exit')) {
-    uopz_allow_exit(true);
-}
